# packages/rest/src/generated/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Credo REST API",
		"version": "0.10.2",
		"description": "Rest API for using Credo over HTTP",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {}
	},
	"paths": {
...
		"/didcomm/proofs/{proofExchangeId}/accept-request": {
			"post": {
				"operationId": "AcceptRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"protocolVersion": "v2",
											"role": "prover",
											"state": "proposal-sent",
											"threadId": "0019d466-5eea-4269-8c40-031b4896c5b7",
											"connectionId": "2aecf74c-3073-4f98-9acb-92415d096834",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"autoAcceptProof": "always",
											"type": "ProofRecord"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a presentation request as prover by sending an accept request message\nto the connection associated with the proof record.",
				"tags": [
					"DIDComm Proofs"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "proofExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DidCommProofsAcceptRequestOptions"
							}
						}
					}
				}
			}
		},
...
		"/didcomm/credentials/{credentialExchangeId}/accept-request": {
			"post": {
				"operationId": "AcceptRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"credentials": [],
											"type": "CredentialRecord",
											"role": "holder",
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"state": "offer-sent",
											"connectionId": "ac6d0fdd-0db8-4f52-8a3d-de7ff8ddc14b",
											"threadId": "82701488-b43c-4d7b-9244-4bb204a7ae26",
											"credentialAttributes": [],
											"protocolVersion": "v1"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a credential request as issuer by sending an accept request message\nto the connection associated with the credential exchange record.",
				"tags": [
					"DIDComm Credentials"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "credentialExchangeId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AcceptCredentialRequestOptions"
							}
						}
					}
				}
			}
		},
...
		"/didcomm/connections": {
			"get": {
				"operationId": "FindConnectionsByQuery",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/DidCommConnectionRecord"
									},
									"type": "array"
								},
								"examples": {
									"Example 1": {
										"value": [
											{
												"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
												"type": "ConnectionRecord",
												"createdAt": "2022-01-01T00:00:00.000Z",
												"did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
												"state": "invitation-sent",
												"role": "responder",
												"invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
												"outOfBandId": "edbc89fe-785f-4774-a288-46012486881d"
											}
										]
									}
								}
							}
						}
					}
				},
				"description": "Find connection record by query",
				"tags": [
					"DIDComm Connections"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "outOfBandId",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					},
					{
						"in": "query",
						"name": "alias",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/DidCommDidExchangeState"
						}
					},
					{
						"in": "query",
						"name": "did",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Did"
						}
					},
					{
						"in": "query",
						"name": "theirDid",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Did"
						}
					},
					{
						"in": "query",
						"name": "theirLabel",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "before",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "after",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/didcomm/connections/{connectionId}": {
			"get": {
				"operationId": "GetConnectionById",
				"responses": {
					"200": {
						"description": "ConnectionRecord",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"type": "ConnectionRecord",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
											"state": "invitation-sent",
											"role": "responder",
											"invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
											"outOfBandId": "edbc89fe-785f-4774-a288-46012486881d"
										}
									}
								}
							}
						}
					}
				},
				"description": "Retrieve connection record by connection id",
				"tags": [
					"DIDComm Connections"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"description": "Connection identifier",
						"in": "path",
						"name": "connectionId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteConnection",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Deletes a connection record from the connection repository.",
				"tags": [
					"DIDComm Connections"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"description": "Connection identifier",
						"in": "path",
						"name": "connectionId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			}
		},
		"/didcomm/connections/{connectionId}/accept-request": {
			"post": {
				"operationId": "AcceptRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"type": "ConnectionRecord",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
											"state": "invitation-sent",
											"role": "responder",
											"invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
											"outOfBandId": "edbc89fe-785f-4774-a288-46012486881d"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a connection request as inviter by sending a connection response message\nfor the connection with the specified connection id.\n\nThis is not needed when auto accepting of connection is enabled.",
				"tags": [
					"DIDComm Connections"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "connectionId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			}
		},
		"/didcomm/connections/{connectionId}/accept-response": {
			"post": {
				"operationId": "AcceptResponse",
				"responses": {
					"200": {
						"description": "ConnectionRecord",
						"content": {
							"application/json": {
								"schema": {},
								"examples": {
									"Example 1": {
										"value": {
											"id": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e",
											"type": "ConnectionRecord",
											"createdAt": "2022-01-01T00:00:00.000Z",
											"did": "did:peer:1zQmfQh1T3rSqarP2FZ37uKjdQHPKFdVyo2mGiAPHZ8Ep7hv",
											"state": "invitation-sent",
											"role": "responder",
											"invitationDid": "did:peer:2.SeyJzIjoiaHR0cHM6Ly9kYTIzLTg5LTIwLTE2Mi0xNDYubmdyb2suaW8iLCJ0IjoiZGlkLWNvbW11bmljYXRpb24iLCJwcmlvcml0eSI6MCwicmVjaXBpZW50S2V5cyI6WyJkaWQ6a2V5Ono2TWtualg3U1lXRmdHMThCYkNEZHJnemhuQnA0UlhyOGVITHZxQ3FvRXllckxiTiN6Nk1rbmpYN1NZV0ZnRzE4QmJDRGRyZ3pobkJwNFJYcjhlSEx2cUNxb0V5ZXJMYk4iXSwiciI6W119",
											"outOfBandId": "edbc89fe-785f-4774-a288-46012486881d"
										}
									}
								}
							}
						}
					}
				},
				"description": "Accept a connection response as invitee by sending a trust ping message\nfor the connection with the specified connection id.\n\nThis is not needed when auto accepting of connection is enabled.",
				"tags": [
					"DIDComm Connections"
				],
				"security": [
					{
						"tenants": [
							"tenant"
						]
					}
				],
				"parameters": [
					{
						"description": "Connection identifier",
						"in": "path",
						"name": "connectionId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/RecordId"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
	"components": {
...
		"schemas": {
...
			"RecordId": {
				"type": "string",
				"example": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e"
			},
...
			"ThreadId": {
				"type": "string",
				"example": "ea4e5e69-fc04-465a-90d2-9f8ff78aa71d"
			},
...
			"DidCommAutoAcceptProof": {
				"description": "Typing of the state for auto acceptance",
				"enum": [
					"always",
					"contentApproved",
					"never"
				],
				"type": "string"
			},
...
			"Record_string.string-or-number_": {
				"properties": {},
				"additionalProperties": {
					"anyOf": [
						{
							"type": "string"
						},
						{
							"type": "number",
							"format": "double"
						}
					]
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"AnonCredsClaimRecord": {
				"$ref": "#/components/schemas/Record_string.string-or-number_"
			},
			"AnonCredsCredentialInfo": {
				"properties": {
					"credentialId": {
						"type": "string"
					},
					"attributes": {
						"$ref": "#/components/schemas/AnonCredsClaimRecord"
					},
					"schemaId": {
						"type": "string"
					},
					"credentialDefinitionId": {
						"type": "string"
					},
					"revocationRegistryId": {
						"type": "string",
						"nullable": true
					},
					"credentialRevocationId": {
						"type": "string",
						"nullable": true
					},
					"methodName": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"linkSecretId": {
						"type": "string"
					}
				},
				"required": [
					"credentialId",
					"attributes",
					"schemaId",
					"credentialDefinitionId",
					"revocationRegistryId",
					"credentialRevocationId",
					"methodName",
					"createdAt",
					"updatedAt",
					"linkSecretId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AnonCredsRequestedAttributeMatch": {
				"properties": {
					"credentialId": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"revealed": {
						"type": "boolean"
					},
					"credentialInfo": {
						"$ref": "#/components/schemas/AnonCredsCredentialInfo"
					},
					"revoked": {
						"type": "boolean"
					}
				},
				"required": [
					"credentialId",
					"revealed",
					"credentialInfo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.AnonCredsRequestedAttributeMatch_": {
				"properties": {},
				"additionalProperties": {
					"$ref": "#/components/schemas/AnonCredsRequestedAttributeMatch"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"AnonCredsRequestedPredicateMatch": {
				"properties": {
					"credentialId": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"credentialInfo": {
						"$ref": "#/components/schemas/AnonCredsCredentialInfo"
					},
					"revoked": {
						"type": "boolean"
					}
				},
				"required": [
					"credentialId",
					"credentialInfo"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Record_string.AnonCredsRequestedPredicateMatch_": {
				"properties": {},
				"additionalProperties": {
					"$ref": "#/components/schemas/AnonCredsRequestedPredicateMatch"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"Record_string.string_": {
				"properties": {},
				"additionalProperties": {
					"type": "string"
				},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"AnonCredsSelectedCredentials": {
				"properties": {
					"attributes": {
						"$ref": "#/components/schemas/Record_string.AnonCredsRequestedAttributeMatch_"
					},
					"predicates": {
						"$ref": "#/components/schemas/Record_string.AnonCredsRequestedPredicateMatch_"
					},
					"selfAttestedAttributes": {
						"$ref": "#/components/schemas/Record_string.string_"
					}
				},
				"required": [
					"attributes",
					"predicates",
					"selfAttestedAttributes"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DidCommProofsAcceptRequestOptions": {
				"properties": {
					"autoAcceptProof": {
						"$ref": "#/components/schemas/DidCommAutoAcceptProof"
					},
					"comment": {
						"type": "string"
					},
					"goalCode": {
						"type": "string",
						"description": "Will be ignored for v1 protocol as it is not supported"
					},
					"goal": {
						"type": "string",
						"description": "Will be ignored for v1 protocol as it is not supported"
					},
					"proofExchangeRecordId": {
						"type": "string"
					},
					"willConfirm": {
						"type": "boolean",
						"default": true
					},
					"useReturnRoute": {
						"type": "boolean",
						"description": "whether to enable return routing on the send presentation message. This value only\nhas an effect for connectionless exchanges."
					},
					"proofFormats": {
						"properties": {
							"anoncreds": {
								"$ref": "#/components/schemas/AnonCredsSelectedCredentials"
							},
							"indy": {
								"$ref": "#/components/schemas/AnonCredsSelectedCredentials"
							}
						},
						"type": "object"
					}
				},
				"required": [
					"proofExchangeRecordId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"DidCommHandshakeProtocol": {
				"description": "Enum values should be sorted based on order of preference. Values will be\nincluded in this order when creating out of band invitations.",
				"enum": [
					"https://didcomm.org/didexchange/1.x",
					"https://didcomm.org/connections/1.x"
				],
				"type": "string"
			},
...
			"Did": {
				"type": "string"
			},
...
			"DidCommAutoAcceptCredential": {
				"description": "Typing of the state for auto acceptance",
				"enum": [
					"always",
					"contentApproved",
					"never"
				],
				"type": "string"
			},
...
			"Record_string.never_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"AnonCredsDidCommAcceptRequestFormat": {
				"$ref": "#/components/schemas/Record_string.never_",
				"description": "This defines the module payload for calling CredentialsApi.acceptRequest. No options are available for this\nmethod, so it's an empty object"
			},
			"DidCommCredentialFormatPayload_CredentialFormats.acceptRequest_": {
				"properties": {
					"indy": {
						"$ref": "#/components/schemas/AnonCredsDidCommAcceptRequestFormat"
					},
					"anoncreds": {
						"$ref": "#/components/schemas/AnonCredsDidCommAcceptRequestFormat"
					}
				},
				"type": "object",
				"description": "Get the payload for a specific method from a list of CredentialFormat interfaces and a method"
			},
			"AcceptCredentialRequestOptions": {
				"properties": {
					"credentialFormats": {
						"$ref": "#/components/schemas/DidCommCredentialFormatPayload_CredentialFormats.acceptRequest_"
					},
					"autoAcceptCredential": {
						"$ref": "#/components/schemas/DidCommAutoAcceptCredential"
					},
					"comment": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"DidCommDidExchangeState": {
				"description": "Connection states as defined in RFC 0023.",
				"enum": [
					"start",
					"invitation-sent",
					"invitation-received",
					"request-sent",
					"request-received",
					"response-sent",
					"response-received",
					"abandoned",
					"completed"
				],
				"type": "string"
			},
			"DidCommDidExchangeRole": {
				"enum": [
					"requester",
					"responder"
				],
				"type": "string"
			},
			"DidCommConnectionType": {
				"enum": [
					"mediator"
				],
				"type": "string"
			},
			"DidCommConnectionRecord": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/RecordId"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"type": {
						"type": "string"
					},
					"did": {
						"$ref": "#/components/schemas/Did"
					},
					"theirDid": {
						"$ref": "#/components/schemas/Did"
					},
					"theirLabel": {
						"type": "string"
					},
					"state": {
						"$ref": "#/components/schemas/DidCommDidExchangeState"
					},
					"role": {
						"$ref": "#/components/schemas/DidCommDidExchangeRole"
					},
					"alias": {
						"type": "string"
					},
					"autoAcceptConnection": {
						"type": "boolean"
					},
					"threadId": {
						"$ref": "#/components/schemas/ThreadId"
					},
					"imageUrl": {
						"type": "string"
					},
					"mediatorId": {
						"type": "string"
					},
					"errorMessage": {
						"type": "string"
					},
					"protocol": {
						"$ref": "#/components/schemas/DidCommHandshakeProtocol"
					},
					"outOfBandId": {
						"type": "string"
					},
					"invitationDid": {
						"$ref": "#/components/schemas/Did"
					},
					"connectionTypes": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/DidCommConnectionType"
								},
								{
									"type": "string"
								}
							]
						},
						"type": "array"
					},
					"previousDids": {
						"items": {
							"$ref": "#/components/schemas/Did"
						},
						"type": "array"
					},
					"previousTheirDids": {
						"items": {
							"$ref": "#/components/schemas/Did"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"createdAt",
					"type",
					"state",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"tenants": {
				"type": "apiKey",
				"name": "x-tenant-id",
				"in": "header",
				"description": "Optional header to set the context to a specific tenant. Only needed when the --multi-tenant feature is enabled. Default value is 'default'"
			}
		}
	},
	"security": [
		{
			"tenants": []
		}
	]
}